backend/requirements.txt
# Core API
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
python-dotenv
python-multipart
requests

# Auth / Security
python-jose[cryptography]
passlib[argon2]
bcrypt

# ML / AI (Relaxed versions to allow resolution)
torch
transformers<4.46.0
sentence-transformers
scikit-learn==1.5.0
numpy<2.0.0
faiss-cpu
pypdf
